// src/interfaces/TickerList.ts
export interface TickerList {
    symbol: string;
    logo: string;
    industry: string;
    marketCap: number;
    weeklyDelta: number;
    yearlyDelta: number;
    exchange: string;
    dailyDelta: number;
    price: number;
    pe: number;
    pb: number;
    roe: number;
    companyName: string;
}